requirements.txt
Django>=5.2.1
python-dotenv>=1.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
Pillow>=10.2.0
djangorestframework-simplejwt>=5.3.1
cryptography>=42.0.0
python-jose>=3.3.0
django-storages>=1.14.2 